# Make fetch() decompress gzip/deflate faster by adding libdeflate