--- a/.gitmodules
+++ b/.gitmodules
@@ -82,3 +82,6 @@ url = https://github.com/oven-sh/zig
 depth = 1
 shallow = true
 fetchRecurseSubmodules = false
+[submodule "src/deps/libdeflate"]
+	path = src/deps/libdeflate
+	url = https://github.com/ebiggers/libdeflate