--- a/packages/bun-types/bun.d.ts
+++ b/packages/bun-types/bun.d.ts
@@ -3100,6 +3100,10 @@ declare module "bun" {
    */
   function openInEditor(path: string, options?: EditorOptions): void;
 
+  const fetch: typeof globalThis.fetch & {
+    preconnect(url: string): void;
+  };
+
   interface EditorOptions {
     editor?: "vscode" | "subl";
     line?: number;