--- a/src/bun.js/javascript.zig
+++ b/src/bun.js/javascript.zig
@@ -2398,6 +2398,9 @@ pub const VirtualMachine = struct {
     // TODO:
     pub fn deinit(this: *VirtualMachine) void {
         this.source_mappings.deinit();
+        if (this.rare_data) |rare_data| {
+            rare_data.deinit();
+        }
         this.has_terminated = true;
     }