### What does this PR do?

TODO: rebase after #12698 is merged. This shares the branch.

Make fetch() decompress gzip/deflate faster by adding libdeflate

```ts
cpu: Apple M3 Max
runtime: bun 1.1.21 (arm64-darwin)

benchmark       time (avg)             (min … max)       p75       p99      p995
-------------------------------------------------- -----------------------------
• roundtrip - "Hello World!".repeat(9999))
-------------------------------------------------- -----------------------------
zlib        172.75 µs/iter   (149.04 µs … 1.21 ms) 174.83 µs 218.38 µs 245.13 µs
libdeflate  136.28 µs/iter   (119.17 µs … 1.49 ms) 136.83 µs    162 µs 175.54 µs

summary for roundtrip - "Hello World!".repeat(9999))
  libdeflate
   1.27x faster than zlib

• gzipSync("Hello World!".repeat(9999)))
-------------------------------------------------- -----------------------------
zlib        127.84 µs/iter (113.54 µs … 219.92 µs) 128.79 µs 150.13 µs 154.88 µs
libdeflate  107.14 µs/iter  (94.79 µs … 197.29 µs) 108.04 µs    134 µs 138.63 µs

summary for gzipSync("Hello World!".repeat(9999)))
  libdeflate
   1.19x faster than zlib

• gunzipSync("Hello World!".repeat(9999)))
-------------------------------------------------- -----------------------------
zlib          42.8 µs/iter     (36 µs … 719.42 µs)  44.08 µs  56.79 µs  63.71 µs
libdeflate   29.44 µs/iter    (24.17 µs … 5.36 ms)  30.46 µs  39.79 µs  45.79 µs

summary for gunzipSync("Hello World!".repeat(9999)))
  libdeflate
   1.45x faster than zlib
```

WIP x64 benchmark



### How did you verify your code works?