@Jarred-Sumner 2024-07-21
x64 is slower!


High-end server:
```zig
cpu: 13th Gen Intel(R) Core(TM) i9-13900
runtime: bun 1.1.21 (x64-linux)

benchmark       time (avg)             (min … max)       p75       p99      p999
-------------------------------------------------- -----------------------------
• roundtrip - "Hello World!".repeat(9999))
-------------------------------------------------- -----------------------------
zlib           171 µs/iter       (151 µs … 776 µs)    170 µs    357 µs    681 µs
libdeflate     151 µs/iter       (146 µs … 794 µs)    147 µs    302 µs    647 µs

summary for roundtrip - "Hello World!".repeat(9999))
  libdeflate
   1.13x faster than zlib

• gzipSync("Hello World!".repeat(9999)))
-------------------------------------------------- -----------------------------
zlib           135 µs/iter       (119 µs … 349 µs)    122 µs    305 µs    336 µs
libdeflate     108 µs/iter    (92'617 ns … 307 µs) 93'492 ns    273 µs    300 µs

summary for gzipSync("Hello World!".repeat(9999)))
  libdeflate
   1.25x faster than zlib

• gunzipSync("Hello World!".repeat(9999)))
-------------------------------------------------- -----------------------------
zlib        32'579 ns/iter    (30'843 ns … 931 µs) 31'205 ns 36'758 ns    466 µs
libdeflate  55'972 ns/iter  (52'675 ns … 1'258 µs) 53'187 ns 82'185 ns    534 µs

summary for gunzipSync("Hello World!".repeat(9999)))
  zlib
   1.72x faster than libdeflate
```

Medium-end desktop:
```zig
cpu: AMD Ryzen 5 3600 6-Core Processor
runtime: bun 1.1.21 (x64-linux)

benchmark       time (avg)             (min … max)       p75       p99      p995
-------------------------------------------------- -----------------------------
• roundtrip - "Hello World!".repeat(9999))
-------------------------------------------------- -----------------------------
zlib        263.83 µs/iter    (237.8 µs … 2.58 ms)  289.6 µs 303.77 µs 309.44 µs
libdeflate  452.38 µs/iter    (243.96 µs … 3.4 ms) 526.56 µs 587.03 µs  599.9 µs

summary for roundtrip - "Hello World!".repeat(9999))
  zlib
   1.71x faster than libdeflate

• gzipSync("Hello World!".repeat(9999)))
-------------------------------------------------- -----------------------------
zlib        196.36 µs/iter (192.85 µs … 221.57 µs) 198.84 µs 208.26 µs 210.57 µs
libdeflate  388.31 µs/iter (177.48 µs … 485.36 µs) 462.84 µs 472.75 µs  473.6 µs

summary for gzipSync("Hello World!".repeat(9999)))
  zlib
   1.98x faster than libdeflate

• gunzipSync("Hello World!".repeat(9999)))
-------------------------------------------------- -----------------------------
zlib         49.05 µs/iter     (44.09 µs … 1.2 ms)  44.92 µs  98.04 µs 102.91 µs
libdeflate   74.44 µs/iter   (65.81 µs … 11.82 ms)  66.82 µs  124.9 µs  131.5 µs

summary for gunzipSync("Hello World!".repeat(9999)))
  zlib
   1.52x faster than libdeflate
```